Отправляю на вторую проверку

+++ controllers/cards.js
@@ -1,48 +1,44 @@
 const Card = require('../models/card');
+const ForbiddenError = require('../errors/forbidden-error');
+const NotFoundError = require('../errors/not-found-error');
 
 // возвращает все карточки
-const getCard = (req, res) => {
+const getCard = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       res.status(200).send(cards);
     })
-    .catch(() => res.status(500).send({ message: 'некорректный запрос к серверу' }));
+    .catch(next);
 };
 
 // создаёт карточку
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
   Card.create({ name, link, owner })
     .then((card) => res.status(200).send(card))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Переданы некоректные данные при создании карточки' });
-      }
-      return res.status(500).send({ message: 'Ошибка на сервере' });
-    });
+    .catch(next);
 };
 
 // удаляет карточку по идентификатору
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(new Error('NotFound'))
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .orFail(() => next(new NotFoundError('Карточка с указанным id не найдена')))
     .then((card) => {
-      res.status(200).send(card);
-    })
-    .catch((err) => {
-      console.log(err.name);
-      if (err.name === 'CastError') {
-        res.status(400).send({ message: 'некоректный запрос' });
-      } if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'карточка с указанным id не найдена' });
+      if (card.owner.toString() !== req.user._id) {
+        next(new ForbiddenError('Нет прав для удаления карточки'));
+      } else {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => {
+            res.status(200).send(card);
+          });
       }
-      return res.status(500).send({ message: 'Ошибка на сервере' });
-    });
+    })
+    .catch(next);
 };
 
 // обновляет лайки карточке
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   const { cardId } = req.params;
   Card.findByIdAndUpdate(cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -52,19 +48,10 @@ const likeCard = (req, res) => {
       res.status(200).send(card);
     })
     // eslint-disable-next-line consistent-return
-    .catch((err) => {
-      console.log(err.name);
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданны некоректные данные для постановки/снятия лайка' });
-      } if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'карточка с указанным id не найдена' });
-      } else {
-        res.status(500).send({ message: 'Ошибка на сервере' });
-      }
-    });
+    .catch(next);
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   const { cardId } = req.params;
   Card.findByIdAndUpdate(cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -74,16 +61,7 @@ const dislikeCard = (req, res) => {
       res.status(200).send(card);
     })
     // eslint-disable-next-line consistent-return
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданны некоректные данные для постановки/снятия лайка' });
-      }
-      if (err.message === 'NotFound') {
-        res.status(404).send({ message: 'карточка с указанным id не найдена' });
-      } else {
-        res.status(500).send({ message: 'Ошибка на сервере' });
-      }
-    });
+    .catch(next);
 };
 
 module.exports = {
